refactor: `CardUiState`를 순수한 데이터 홀더로 정상화
