+++ app/src/main/java/woowacourse/payments/ui/features/addcard/CardUiState.kt
@@ -1,8 +1,9 @@
 package woowacourse.payments.ui.features.addcard
 
-import woowacourse.payments.domain.Card
-import woowacourse.payments.ui.features.addcard.components.ExpireDateStatus
-import java.time.YearMonth
+import woowacourse.payments.domain.Card.Companion.MAX_LENGTH_CARD_NUMBER
+import woowacourse.payments.domain.Card.Companion.MAX_LENGTH_EXPIRE_DATE
+import woowacourse.payments.domain.Card.Companion.MAX_LENGTH_OWNER_NAME
+import woowacourse.payments.domain.Card.Companion.MAX_LENGTH_PASSWORD
 
 data class CardUiState(
     val cardNumber: String = "",
@@ -10,29 +11,6 @@ data class CardUiState(
     val ownerName: String = "",
     val password: String = "",
 ) {
-    val isValidCardNumber: Boolean
-        get() = cardNumber.length == MAX_LENGTH_CARD_NUMBER && cardNumber.all { it.isDigit() }
-
-    val isValidExpireDate: Boolean
-        get() = expireDateStatus == ExpireDateStatus.Valid
-
-    val expireDateStatus: ExpireDateStatus
-        get() {
-            if (expireDate.length != MAX_LENGTH_EXPIRE_DATE) return ExpireDateStatus.Typing
-            val mm = expireDate.substring(0, 2).toIntOrNull() ?: return ExpireDateStatus.Invalid
-            val yy = expireDate.substring(2, 4).toIntOrNull() ?: return ExpireDateStatus.Invalid
-            if (mm !in 1..12) return ExpireDateStatus.InvalidMonth
-            val year = 2000 + yy
-            val ym = YearMonth.of(year, mm)
-            return if (ym >= YearMonth.now()) ExpireDateStatus.Valid else ExpireDateStatus.Expired
-        }
-
-    val isValidPassword: Boolean
-        get() = password.length == MAX_LENGTH_PASSWORD && password.all { it.isDigit() }
-
-    val isFormValid: Boolean
-        get() = isValidCardNumber && isValidExpireDate && isValidPassword
-
     fun withCardNumber(newNumber: String): CardUiState {
         if (newNumber.length <= MAX_LENGTH_CARD_NUMBER && newNumber.all { it.isDigit() }) {
             return this.copy(cardNumber = newNumber)
@@ -59,26 +37,5 @@ data class CardUiState(
             return this.copy(password = newPassword)
         }
         return this
-    }
-
-    fun toDomainCard(): Card? {
-        if (!isFormValid) return null
-
-        val mm = expireDate.substring(0, 2).toInt()
-        val yy = expireDate.substring(2, 4).toInt()
-
-        return Card(
-            cardNumber = this.cardNumber,
-            expireDate = YearMonth.of(2000 + yy, mm),
-            ownerName = this.ownerName.ifEmpty { null },
-            password = this.password,
-        )
-    }
-
-    companion object {
-        private const val MAX_LENGTH_CARD_NUMBER = 16
-        const val MAX_LENGTH_EXPIRE_DATE = 4
-        const val MAX_LENGTH_OWNER_NAME = 30
-        private const val MAX_LENGTH_PASSWORD = 4
     }
 }
